Distinguish error messages for many maps and no maps
Before, a common error message was displayed in the case of both no
map being available for a value and many being available. To correctly
know the reason, the user will have to go back and check manually with
clues from the log which is tedioius. The error message has been
updated to distinguish the two possibilities and make life easier.