Fix None tag bug that led to failed path building
The dict.update() is used to update the tags in the file. The
dict.update() method does not return anything, rather it makes changes
directly to the object. But in the implementation, it was expecting
the return, due to which the tags were set to none. Because of this,
path building failed when attempted to get modified path. This has
been fixed in this commit.