Fix api reference doc not showing on git pages
The api reference page of the doc on git pages does not show up. I
believe the reason for this might be because almirah is not installed
in the environment while building docs in git. Since almirah api
reference uses docstrings from the code base, maybe its presence would
fix this. To this end, almirah is added to the requirements in
environment. Now that almirah is available on pip, the environment can
download stable release from there. This way docs has a stable version
too.