Fix sensitive data field leakage in nirs data conversion
The format conversion pipelines serve two purposes:

1) to convert to formats that are more openly support by readily
available toolkits that help analyse the data

2) to anonymize sensitive info before public release

The nirs modality conversion which depended on mne-nirs did not have
anonymization already in the pipeline, due to which the subject name
among other details were leaking out. This is not desired. To stop
this, anonymization is added using mne's Raw object method anonymize
that inturn invokes anonymize_info.