Reflect dataset name change and update dataset docs
The official name for the dataset has been decided. It will be called
the CALM-BRAINS dataset. To reflect, this file path have been changed.

Tutorials on how to access, use, and refer to the dataset have been
updated too. The examples are barebones. Hopefully in the future we
will expand on them and also add more to the collection.