Update gh-action workflows to use poetry for consistency
The project uses two actions: one to publish docs and the other to
release on PyPI. The formatting used for both has been changed to make
them consistent.

Further, the publish docs has been update to use a poetry based flow
to maintain consistency with local development and testing. Before,
the action ammaraskar/sphinx-action was used and the requirements
provided in a separate file. This in turn hinged on the availability
of the latest version of the project on PyPI for download which might
lead to documentation inconsistencies in API reference. Now, the whole
process depends on the poetry environment used in development. Hence,
maintaining consistency.