Group functions in utility module based on relevance
The utility module has grown in size and has a lot of miscellaneous
things. All of them necessary, ranging from hiding dirty work, to
doing repetitive tasks. But, instead of big file, a collection of
smaller files will make it easy to work with and scale later. This
commit does this. At some point in the future, File-conversion
functionality needs refactoring.