Add ability to uniquify ORM Models
It is important that ORM models are unique, namely, changes to one
object representing an entity is reflected in another object
representing the same entity. This is important to maintain
consistency. The added uniquify module supports this.

It provides a class decorator that ensures uniqueness in all cases, by
maintaining a cache as well as checking index if not in cache. This
adds a bit of an overhead to object creation which hopefully will not
be much. If so, alternatives will have to be explored.

The rough flow for object creation after this will be like:

1. Provide same arguments to create object
2. If all required arguments found, get identifiers and create
object. This is to ensure all arguments needed are present.
3. Post creation, if object found in index take that. This is to
ensure if other non-identifier values such as details in Specification
are same as in index.
4. Add to cache if not present and if present choose that version.

Based on usage, changes might be needed to ensure decent speed and
functionality.