Change docs config to accomodate uniquified classes
When generating documentation for class with decorators that override
new, documentation for the class is missing and sphinx warns that
maximum recursion depth was exceeded in default conf configuration
settings. This is because `__init__` and `__new__` are kind of
special methods and have special handling by the ModuleAnalyzer. So,
the sphinx codebase does not accomodate these by default.

To work around these, changes are needed to how autodoc is processed:

 - Exclude `__init__` and `__new__`
 - Seperate class signature from `__init__`

Reference: sphinx-doc/sphinx#12276