Add ability to move Layout and its entries abstractly
The index stores absolute paths. So, say you share your dataset with
your friend and indexing is an expensive operation, or you have packed
extra info into your index using `tag_funcs`, then it is preferrable
to share you index with them too. This reduces them performing similar
tasks and reusing your efforts. But but, how will they have the same
absolute path. Exactly, so we need to change the Layout locations for
the index to be of help. `Layout.move_root` accomplishes this.