Fix bug to prevent errors on indexing metadata when json not present
If the metadata argument is set during indexing, almirah attempts to
find the related json file with the same filename and add its contents
of nestedness one. But if no json is found, then an error is
thrown. Instead, it would better to skip it as all files might not
have metadata json files. This commit fixes this.