Add request mode to Database
Say 1000 people want to access a database for read-only purpose, it
might not be the best idea to create accounts for everyone. Instead
would it be possible to just create one account, and use it to serve
content. No, not share the same credentials with everyone, rather
setup a service in the host that all people can request to. The
service using this one account in the backend to serve content. The
service can use almirah to retrieve records and pass it on as a HTTP
response too :)

If a url endpoint is associated with the service, then HTTP requests
can be directed here. This is what request mode does. To work in
request mode, provide backend='request'.