Fix bug in DBManager metadata reflection that lead to metadata=None
On creation, the DBmanager reflects the schema of the database it
connects to retreive the structure of all existing table. This is
implemented using `MetaData.reflect()` of SQLAlchemy. This method
returns none and populated a `MetaData` instance in place. Due to
oversight, the return of this method was assigned to `metadata` which
caused the bug. This has been fixed.