Add authentication support for Database in request mode
Request mode support for Database was adding in commit
7611880b6ab8e89e2cdd5e8ac6fcc97e56bc4709. The main reasoning behind
this was security and scalability without direct access to the
database. But there is not way to authenticate, so anyone who knows
the endpoint can send a valid request and retrieve data. This is not
desirable.

To prevent this, authentication is added. Using `connect()` in
request-mode raises a request to the {host}/authenticate. This
endpoint should return a token or authorization header that can act as
a bearer. This will be passed during query request to the host
endpoint. The user is responsible for setting up the endpoint. A
tutorial on this will follow at some point.