Fix bug in replace transform that raised errors
The replace transform is used to switch values into the source
records. This is performed either by providing an external mapping for
all values or by providing specific values and their to's. A key to
this running smoothly in the case of external mapping is for the index
to match the values in the series. Since was not the case, it raise
lead to empty records as strict was set or errors. This has been
fixed. In this process, the replace column config keys have been
updated to be consistent across both replace strategies.