Add support for Date type in mapping config
The mapping config is used to mgirate records from one db to
another. Datetime is not always the best way to represent dates as it
appends them with '00:00:00' if no time is provided. This can be the
case in columns representing for example date of birth. This is not
desirable. For this, support 'Date' type is introduced. This column
type can be used in the mapping config by specifying `dtype: date`.