Fix bug in finding common rows that lead to empty records
Common rows between two tables is required for example while writing
to prevent inserting a record twice. A mask where `True` represents a
records existance in both child and parent is returned
`common_rows`. This is calculated by performing a merge. For the mask
to be functional, the index of the mask and the child dataframe should
be the same. Since this was not case, it lead to empty records. This
commit fixes this.