Fix bug due to premature closing of file in reading multi yaml files
Configs such as mapping consist of multiple yamls in a single
file. These are read using `read_multi_yaml`. Since `safe_load_all` of
pyyaml returns generators, it is necessary to keep the file connection
open until all yamls are read. Since this was not the case, an error
was raised. As a fix, all yamls are read and stored in a list and
returned.