Fix broken import in tutorial docs
In the update from version 0.2.0 to 0.3.0, utility functions which
were monolithically stores in utils.py were broken down into separate
submodules for ease of management. In this process,
`get_tutorial_dataset` was moved from utils.py to utils/lib.py. Since
this change was not reflected in docs nor the new import location
added to __init__, the import in tutorial was broken. Adding the new
import location to utils/__init__.py fixes this and resolves #1.