Move tutorial dataset from test to inside package
The tutorial dataset was stored in tests/data/tutorial and a utility
function copied it to provided part. But it seems like this tests
directory is included while packaging, due to which an error is thrown
as raised issue #1. To fix this, tutorial dataset is created using a
series of python Path statements instead of copying a dummy file
directory.