Fix unintended type changes during db record retrieval
Pandas is used to retrieve records from a database. Due to the default
behavior of pandas, column of int for example are converted to float
in the presence of `NA` values. This is not desirable as say boolean
is stored as TINYINT in mysql and integers 1, 0, and `NA` are present,
these become 1.0, 0.0, `NA`. To prevent this, appropriate pandas type
is inferred by reflecting table schema and this dtype is enforced.