Add read-only gsheet mode to Database
This commit adds the ability to read data from google sheets. Here the
google sheet is treated as a Database and each sheet a table. For this
to function, a JSON keyfile is required.

With this commit, it should now be possible to migrate data from
google sheets to SQL tables as well.