Update outdated tutorials to be compatible with version 0.3.0
The tutorials have not been updated since v0.1.0, due to this
breakages as mentioned in #1 have occured. To fix this and other
unnoticed things, the tutorial and user guide have been revamped and
tested locally. New additions have been incorporated. There is still
work to do for a more friendly package and help is welcome for this.