Fix bug in Database token property that lead to None always
The token property in Database checks if a token has been assigned on
connecting to request mode hosts and retrieves it. But there was no
return statement and hence None was returned. When token is not
available due to non-connected status of Database, a premature error
was thrown. Both of these have been fixed by providing necessary
return statement and default value to getattr respectively.