Add ability to skip entries while indexing Layouts
To cover usecases where one needs to skip certain entries, the skip
parameter has been added to the index() method. skip takes in a list
of string and entries that regex-match any of the given values are
skipped. skip can be used either to ignore bad entries or to speed up
indexing by not diving into huge unnecessary directories.