Fix bug in format converters
Over the releases, the build_modified_path() method of File was
removed as it did not align with the almirah concept. Path building is
a function of a Specification and not File. As a side effect, the
converter function broke down but went unnoticed. This commit fixes
them by mimicing build_modified_path() using Specification.build_path()
with strict parameter set to False.

To retrieve specification, instead of passing destination as string,
the destination is passed as a Layout. This allows building paths
using tags. Tags of source are modified as necessary in conversion
routines.