Fix type bug in column type conversion
Tables are represented using pandas DataFrames in almirah. While
retrieving table records from sources, their natural type is changed
in many scenarios - while retrieving from request mode endpoint, or
while retrieving from other sources, or while transforming. To
preserve it, they are convered back to a type as close to the natural
type as possible.

A recent commit introduced initial retrieval of all records as string
before type conversion. This introduced a bug in column type
conversion as an error was thrown on attempting to insert non-string
types into the column. This commit fixes this using a stop-gap
solution. This solution will have to revisited for a more elegant
solution.