Add ability to pad retrieved tags while organizing
Tag are used to build paths. And path are strings. So, though
numerically 99 > 100, as they are strings, after path building 100 is
listed before 99. This might get confusing when dealing with numbers
of irregular digit count. For this purpose, padding is introduced as
an organizational rule. It can accept the length to pad for and the
character to use for padding as arguments.