Update links broken due to ownership transfer
In the interest of long term maintainance of the repo, the ownership
has been transferred to the lab organization. This will enable future
members of the lab to continue contributing and using the tool to
maintain datasets. But due to this, links in the docs, and readme were
broken and non-functional. This commit fixes this.